Backend/requirements.txt
flask
flask-cors
werkzeug
python-dotenv
cryptography